+++ pnet-data-api-java/src/main/java/at/porscheinformatik/happyrest/AbstractRestAttribute.java
@@ -1,5 +1,7 @@
 package at.porscheinformatik.happyrest;
 
+import java.util.Objects;
+
 /**
  * @author yda
  */
@@ -26,6 +28,31 @@ public abstract class AbstractRestAttribute implements RestAttribute
     public Object getValue()
     {
         return value;
+    }
+
+    @Override
+    public int hashCode()
+    {
+        return Objects.hash(name, value);
+    }
+
+    @Override
+    public boolean equals(Object obj)
+    {
+        if (this == obj)
+        {
+            return true;
+        }
+        if (obj == null)
+        {
+            return false;
+        }
+        if (getClass() != obj.getClass())
+        {
+            return false;
+        }
+        AbstractRestAttribute other = (AbstractRestAttribute) obj;
+        return Objects.equals(name, other.name) && Objects.equals(value, other.value);
     }
 
     @Override